Resolves: #i119844# Spin effect amount property lost when saving pptx to odp
(cherry picked from commit a7eedd28dd66cec1a668c3df6a8568dcad0f858f)

+++ oox/source/ppt/timenodelistcontext.cxx
@@ -718,7 +718,7 @@ namespace oox { namespace ppt {
                 if(attribs.hasAttribute( XML_by ) )
                 {
                     sal_Int32 nBy = attribs.getInteger( XML_by, 0 );
-                    pNode->setBy( makeAny( nBy ) );
+                    pNode->setBy( makeAny( (double)nBy ) );
                 }
                 if(attribs.hasAttribute( XML_from ) )
                 {
